chore(contextHeader): move component slots to ContextHeader.slots.jsx
