+++ src/components/ContextHeader/ContextHeader.jsx
@@ -6,23 +6,7 @@ import React from 'react';
 import { useSlot } from '../../hooks/useSlot';
 
 import styles from './ContextHeader.module.scss';
-
-export const ContextHeaderTopSection = ({ children }) => <>{children}</>;
-export const ContextHeaderActionsSection = ({ children }) => <>{children}</>;
-
-ContextHeaderActionsSection.propTypes = {
-	children: PropTypes.oneOfType([
-		PropTypes.arrayOf(PropTypes.node),
-		PropTypes.node,
-	]),
-};
-
-ContextHeaderTopSection.propTypes = {
-	children: PropTypes.oneOfType([
-		PropTypes.arrayOf(PropTypes.node),
-		PropTypes.node,
-	]),
-};
+import { ContextHeaderActionsSection, ContextHeaderTopSection } from './ContextHeader.slots';
 
 const cx = classNames.bind(styles);
