+++ src/components/ContextHeader/ContextHeader.stories.mdx
@@ -7,7 +7,8 @@ import {
 	CONTEXT_HEADER_MOCK_BREAD_CRUMB_ITEMS,
 	CONTEXT_HEADER_MOCK_TITLE,
 } from './ContextHeader.mock.js';
-import ContextHeader, { ContextHeaderTopSection, ContextHeaderActionsSection } from './ContextHeader.jsx';
+import ContextHeader from './ContextHeader.jsx';
+import { ContextHeaderActionsSection, ContextHeaderTopSection } from './ContextHeader.slots.jsx';
 
 <Meta title="ContextHeader" component={ContextHeader} />
